requirements.txt
alembic==1.11.1
certifi==2023.5.7
charset-normalizer==3.2.0
conf==0.4.1
greenlet==2.0.2
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
urllib3==2.0.3
